Feature/translate distinct (#42)
- add translation for distinct constraints
- refactor splitting into modules
- update statistics
